Moved changes from feature/socket-refactor to a fresh branch since that branch is outdated as heck and I'm not gonna deal with that rebase nonsense! Contains slight fixes due to its outdated-ness, but none of the requested changes (yet).
