+++ src/web-marbles.js
@@ -27,19 +27,11 @@ physics.world.setGravity(config.physics.gravity);
 // Set up game logic
 const game = require("./server/game");
 
-// Set up gameplay socket
-const sockets = require("./server/network/sockets");
-const socketGameplay = sockets.setupGameplay(db, config, game);
-
-// Set game socketManager
-game.setSocketManager(socketGameplay);
-
 // Set up chat
-let discordManager = null,
-	socketChat;
+let discordManager = null;
 if (config.discord.enabled) {
 	discordManager = require("./server/chat/discord-manager");
-	socketChat = discordManager.initialize(db);
+	discordManager.initialize(db);
 }
 
 // Express connections
@@ -203,7 +195,7 @@ function shutdown() {
 		log.warn("Termination signal received. Shutting down web-marbles...".yellow);
 
 		// Inform any connected clients
-		socketGameplay.emit(JSON.stringify({
+		game.socket.emit(JSON.stringify({
 			content: "The server is shutting down.",
 			classNames: "red exclamation"
 		}));
@@ -239,9 +231,10 @@ function shutdown() {
 		log.warn("PHYSICS stopped");
 
 		// µWebSockets
-		if (socketChat) socketChat.close();
-		socketGameplay.close();
-		sockets.close();
+		if(discordManager)
+			discordManager.socket.close();
+		game.socket.close();
+		require("./server/network/sockets-helper").stopListening();
 		log.warn("µWS server closed");
 
 		// Once promises resolve, we should be done
